+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 - Return bucket to player when concrete bucket is drank & set stack size to 1
 - Disable Shaped Packing & regular crafing recipes in Mechanical Crafters 
 - Buff Sophisticated Backpacks/Chests upgrading recipes
+- Replace create factory logistics with create fluid logstics
+- Replace iron with wrought iron in smelting recipes for ores/smeltery
 
 **Mod Changes**
 - jamlib-forge-1.3.5+1.20.1 -> jamlib-forge-1.3.6+1.20.1-patch.1